pyproject.toml
[project]
name = "gitmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.46",
    "mcp[cli]",
    "python-dotenv",
    "requests",
    "GitPython",
]
